Refine gradle script
Move common parts to root project
